Refactor domain classes by moving them to package `simple.escp.dom`.
